+++ enclave-runtime/Cargo.toml
@@ -127,9 +127,9 @@ its-sidechain = { path = "../sidechain/sidechain-crate", default-features = fals
 sp-io = { optional = true, default-features = false, features = ["disable_oom", "disable_panic_handler", "disable_allocator", "sgx"], path = "../core-primitives/substrate-sgx/sp-io" }
 
 # litentry
-lc-stf-task-initialization = { path = "../litentry/core/stf-task/initialization", default-features = false, features = ["sgx"] }
+lc-stf-task-sender = { path = "../litentry/core/stf-task/sender", default-features = false, features = ["sgx"] }
 litentry-primitives = { path = "../litentry/primitives", default-features = false, features = ["sgx"] }
-lc-stf-task-handler = { path = "../litentry/core/stf-task/handler", default-features = false, features = ["sgx"] }
+lc-stf-task-receiver = { path = "../litentry/core/stf-task/receiver", default-features = false, features = ["sgx"] }
 
 # substrate deps
 frame-support = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.28" }
