+++ litentry/core/stf-task/sender/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "lc-stf-task-initialization"
+name = "lc-stf-task-sender"
 version = "0.1.0"
 authors = ["Litentry dev"]
 edition = "2021"
@@ -35,6 +35,7 @@ itp-extrinsics-factory = { path = "../../../../core-primitives/extrinsics-factor
 
 # litentry
 litentry-primitives = { path = "../../../primitives", default-features = false }
+
 
 [features]
 default = ["std"]
