+++ litentry/core/stf-task/sender/src/lib.rs
@@ -34,6 +34,7 @@ pub mod sgx_reexport_prelude {
 use itp_types::AccountId;
 pub mod error;
 pub mod stf_task_sender;
+// pub mod stf_task_receiver;
 // pub mod https_client;
 pub use error::Result;
